removed modern component
