+++ app/components/wallet/WalletAdd.js
@@ -1,67 +1,61 @@
 // @flow
 import React, { Component } from 'react';
 import { observer } from 'mobx-react';
-import { Button } from 'react-polymorph/lib/components/Button';
-import { ButtonSkin } from 'react-polymorph/lib/skins/simple/ButtonSkin';
-import classnames from 'classnames';
 import { defineMessages, intlShape, FormattedHTMLMessage } from 'react-intl';
-import styles from './WalletAdd.scss';
+
+import SvgInline from 'react-svg-inline';
+import logoIcon from '../../assets/images/yoroi-logo-white.inline.svg';
+import settingsIcon from '../../assets/images/top-bar/setting-active.inline.svg';
+import daedalusIcon from '../../assets/images/top-bar/daedalus-migration.inline.svg';
+
 import { MAX_ADA_WALLETS_COUNT } from '../../config/numbersConfig';
+import styles from './WalletAdd.scss';
 
 const messages = defineMessages({
   title: {
     id: 'wallet.add.dialog.title.label',
     defaultMessage: '!!!Add wallet',
   },
+  subTitle: {
+    id: 'wallet.add.dialog.subtitle.label',
+    defaultMessage: '!!!Yoroi light wallet for Cardano',
+  },
   createDescription: {
     id: 'wallet.add.dialog.create.description',
-    defaultMessage: '!!!Create a new wallet',
+    defaultMessage: '!!!Create wallet',
   },
-  useTrezorDescription: {
-    id: 'wallet.add.dialog.trezor.description',
-    defaultMessage: '!!!Connect to Trezor',
-  },
-  useLedgerDescription: {
-    id: 'wallet.add.dialog.ledger.description',
-    defaultMessage: '!!!Connect to Ledger',
+  connectToHardwareDescription: {
+    id: 'wallet.add.dialog.hardware.description',
+    defaultMessage: '!!!Connect to hardware wallet',
   },
   restoreDescription: {
     id: 'wallet.add.dialog.restore.description',
-    defaultMessage: '!!!Restore wallet from backup',
-  },
-  restorePaperDescription: {
-    id: 'wallet.add.dialog.restore.paper.description',
-    defaultMessage: '!!!Restore Yoroi Paper Wallet',
+    defaultMessage: '!!!Restore wallet',
   },
   restoreNotificationMessage: {
     id: 'wallet.add.dialog.restoreNotificationMessage',
     defaultMessage: '!!!Wallet restoration is currently in progress. Until it completes, it is not possible to restore or import new wallets.',
   },
-  createTrezorWalletNotificationMessage: {
-    id: 'wallet.add.dialog.createTrezorWalletNotificationMessage',
-    defaultMessage: '!!!Trezor Connect is currently in progress. Until it completes, it is not possible to restore or import new wallets.',
+  transferFundsTitle: {
+    id: 'daedalusTransfer.wallet.add.page.link.label',
+    defaultMessage: '!!!Transfer funds from Daedalus wallet to Yoroi',
   },
-  createLedgerWalletNotificationMessage: {
-    id: 'wallet.add.dialog.createLedgerWalletNotificationMessage',
-    defaultMessage: '!!!Ledger Connect is currently in progress. Until it completes, it is not possible to restore or import new wallets.',
-  }
 });
 
 type Props = {
-  onTrezor: Function,
-  isCreateTrezorWalletActive: boolean,
-  onLedger: Function,
-  isCreateLedgerWalletActive: boolean,
   onCreate: Function,
   onRestore: Function,
-  onPaperRestore: Function,
+  onHardwareConnect: Function,
+  onSettings: Function,
+  onDaedalusTransfer: Function,
   isRestoreActive: boolean,
+  title: string,
+  subTitle: string,
   classicTheme: boolean,
-  title: string
 };
 
 @observer
-export default class WalletAdd extends Component<Props> {
+export default class WalletAddModern extends Component<Props> {
   static contextTypes = {
     intl: intlShape.isRequired,
   };
@@ -69,92 +63,80 @@ export default class WalletAdd extends Component<Props> {
   render() {
     const { intl } = this.context;
     const {
-      onTrezor,
-      isCreateTrezorWalletActive,
-      onLedger,
-      isCreateLedgerWalletActive,
       onCreate,
       onRestore,
-      onPaperRestore,
+      onHardwareConnect,
+      onSettings,
+      onDaedalusTransfer,
       isRestoreActive,
-      classicTheme,
-      title
+      title,
     } = this.props;
 
-    const componentClasses = classnames([styles.component, 'WalletAdd']);
-    const createWalletButtonClasses = classnames([
-      classicTheme ? 'primary' : 'outlined',
-      'createWalletButton'
-    ]);
-    const restoreWalletButtonClasses = classnames([
-      classicTheme ? 'primary' : 'outlined',
-      'restoreWalletButton'
-    ]);
-    const buttonsContainerClasses = classnames([
-      classicTheme ? styles.buttonsContainerClassic : styles.buttonsContainer
-    ]);
-
     let activeNotification = null;
-    if (isCreateTrezorWalletActive) {
-      activeNotification = 'createTrezorWalletNotificationMessage';
-    } else if (isCreateLedgerWalletActive) {
-      activeNotification = 'createLedgerWalletNotificationMessage';
-    } else if (isRestoreActive) {
+    if (isRestoreActive) {
       activeNotification = 'restoreNotificationMessage';
     }
 
     return (
-      <div className={componentClasses}>
-        <div className={buttonsContainerClasses}>
-          {!classicTheme && (
-            <div className={styles.title}>{title}</div>
-          )}
-
-          <Button
-            className="primary"
-            label={intl.formatMessage(messages.useLedgerDescription)}
-            onMouseUp={onLedger}
-            skin={ButtonSkin}
-          />
-          <Button
-            className="primary"
-            label={intl.formatMessage(messages.useTrezorDescription)}
-            onMouseUp={onTrezor}
-            skin={ButtonSkin}
-          />
-          <Button
-            className={createWalletButtonClasses}
-            label={intl.formatMessage(messages.createDescription)}
-            onMouseUp={onCreate}
-            skin={ButtonSkin}
-          />
-          <Button
-            className={restoreWalletButtonClasses}
-            label={intl.formatMessage(messages.restoreDescription)}
-            onMouseUp={onRestore}
-            skin={ButtonSkin}
-          />
-          <Button
-            className="primary restorePaperWalletButton"
-            label={intl.formatMessage(messages.restorePaperDescription)}
-            onMouseUp={onPaperRestore}
-            skin={ButtonSkin}
-          />
-          {activeNotification ? (
-            <div className={styles.notification}>
-              <FormattedHTMLMessage
-                {...messages[activeNotification]}
-                values={{ maxWalletsCount: MAX_ADA_WALLETS_COUNT }}
-              />
+      <div className={styles.heroWrapper}>
+        <div className={styles.hero}>
+          <div className={styles.settingsBar}>
+            <button type="button" onClick={onSettings} className={styles.settingsBarLink}>
+              <SvgInline svg={settingsIcon} width="30" height="30" />
+            </button>
+          </div>
+          <div className={styles.heroInner}>
+            <div className={styles.heroLeft}>
+              <SvgInline svg={logoIcon} className={styles.heroLogo} />
+              <h2 className={styles.heroTitle}>{title}</h2>
+              <h3 className={styles.heroSubTitle}>{intl.formatMessage(messages.subTitle)}</h3>
             </div>
-          ) : null}
-
+            <div className={styles.heroRight}>
+              <div className={styles.heroCardsList}>
+                <button type="button" onClick={onHardwareConnect}>
+                  <div className={styles.heroCardsItem}>
+                    <div className={`${styles.heroCardsItemBg} ${styles.connect}`} />
+                    <div className={styles.heroCardsItemTitle}>
+                      {intl.formatMessage(messages.connectToHardwareDescription)}
+                    </div>
+                  </div>
+                </button>
+                <button type="button" onClick={onCreate}>
+                  <div className={styles.heroCardsItem}>
+                    <div className={`${styles.heroCardsItemBg} ${styles.create}`} />
+                    <div className={styles.heroCardsItemTitle}>
+                      {intl.formatMessage(messages.createDescription)}
+                    </div>
+                  </div>
+                </button>
+                <button type="button" onClick={onRestore}>
+                  <div className={styles.heroCardsItem}>
+                    <div className={`${styles.heroCardsItemBg} ${styles.restore}`} />
+                    <div className={styles.heroCardsItemTitle}>
+                      {intl.formatMessage(messages.restoreDescription)}
+                    </div>
+                  </div>
+                </button>
+                {activeNotification ? (
+                  <div className={styles.notification}>
+                    <FormattedHTMLMessage
+                      {...messages[activeNotification]}
+                      values={{ maxWalletsCount: MAX_ADA_WALLETS_COUNT }}
+                    />
+                  </div>
+                ) : null}
+              </div>
+              <button type="button" onClick={onDaedalusTransfer} className={`${styles.heroCardsItem} ${styles.heroCardsItemLink}`}>
+                <SvgInline svg={daedalusIcon} width="45" height="40" className={styles.heroCardsItemLinkIcon} />
+                <div className={styles.heroCardsItemTitle}>
+                  {intl.formatMessage(messages.transferFundsTitle)}
+                </div>
+              </button>
+            </div>
+          </div>
         </div>
-
-        {!classicTheme && <div className={styles.background} />}
-
-        {!classicTheme && <div className={styles.walletImage} />}
       </div>
     );
   }
+
 }
